Use stack.
