+++ .gitignore
@@ -3,3 +3,10 @@ Main
 *.o
 TAGS
 shell*
+*#
+*~
+dist/
+.cabal-sandbox/
+cabal.sandbox.config
+TAGS
+.stack-work/
