+++ scheme.cabal
@@ -1,0 +1,48 @@
+name:                scheme
+version:             0.1.0.0
+homepage:            https://en.wikibooks.org/wiki/Write_Yourself_a_Scheme_in_48_Hours
+license:             BSD3
+license-file:        LICENSE
+author:              drets
+copyright:           2017 drets
+category:            Application
+build-type:          Simple
+extra-source-files:  README.md
+cabal-version:       >=1.10
+
+library
+  hs-source-dirs:      src
+  exposed-modules:     Scheme.Parser, Scheme.Eval, Scheme.Types
+  build-depends:       base >= 4.7 && < 5
+                     , parsec
+                     , HUnit
+                     , vector
+                     , mtl
+  default-language:    Haskell2010
+  ghc-options:         -Wall -fno-warn-unused-do-bind
+
+executable scheme-exe
+  hs-source-dirs:      app
+  main-is:             Main.hs
+  ghc-options:         -Wall -fno-warn-unused-do-bind
+                       -threaded
+  build-depends:       base
+                     , scheme
+  default-language:    Haskell2010
+
+test-suite scheme-test
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test
+  main-is:             Spec.hs
+  build-depends:       base
+                     , parsec
+                     , scheme
+                     , HUnit
+                     , vector
+  ghc-options:         -Wall -fno-warn-unused-do-bind
+                       -threaded
+  default-language:    Haskell2010
+
+source-repository head
+  type:     git
+  location: git@github.com:drets/scheme-haskell-48h.git
