+++ stack.yaml
@@ -1,0 +1,8 @@
+resolver: lts-8.5
+
+packages:
+- '.'
+
+nix:
+  enable: true
+  shell-file: default.nix
